blogilistan laajennus, step9

+++ osa4/blogilista/controllers/blogs.js
@@ -34,8 +34,23 @@ blogsRouter.post('/', async (req, res) => {
 })
 
 blogsRouter.delete('/:id', async (req, res) => {
-    await Blog.findByIdAndDelete(req.params.id)
-    res.status(204).end()
+    const decodedToken = await jwt.verify(req.token, process.env.SECRET)
+    if (!decodedToken.id) {
+        res.status(401).json({ error: 'token invalid' })
+    }
+
+    const user = await User.findById(decodedToken.id)
+    console.log('dekoodattu käyttäjä: ', user.id)
+
+    const blog = await Blog.findById(req.params.id)
+    console.log('blogi: ', blog.user.toString())
+
+    if (blog.user.toString() === user.id) {
+        await Blog.findByIdAndDelete(req.params.id)
+        res.status(204).end()
+    } else {
+        res.status(401).json({ error: 'not access to delete this blog'})
+    }
 })
 
 blogsRouter.put('/:id', async (req, res) => {
